# EDIT - first_block_test.json fix, etc. first_block_test.json의 tx_cbor 관련 내용을 유효한 값으로 설정한 후 출력 테스트까지 완료하였습니다. 또, min 함수에서 type 관련 빌드 에러가 생겨서 int32_t를 uint32_t로 변경하였습니다.

+++ src/plugins/chain_plugin/include/static_merkle_tree.hpp
@@ -51,7 +51,7 @@ public:
   void generate(vector<hash_t> &merkle_contents) {
     const bytes dummy_leaf(32, 0); // for SHA-256
 
-    auto min_addable_size = min(MAX_MERKLE_LEAVES, merkle_contents.size());
+    auto min_addable_size = min(static_cast<uint64_t>(MAX_MERKLE_LEAVES), merkle_contents.size());
 
     for (int i = 0; i < min_addable_size; ++i)
       m_merkle_tree[i] = merkle_contents[i];
